matrix_actions_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
matrix-nio==0.25.2
